requirements.txt
Flask==3.0.2
PyPDF2==3.0.1
azure-ai-inference
azure-core==1.30.0